pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biocheck"
version = "0.1.0"
description = "Standalone validator for biological sequence and structure files"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.81",
    "click>=8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
biocheck = "biocheck.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
